Cleaned up some warnings Fixed an issue with step over when sp was 0 Fixed a bug with goto dialog box Renamed newimportvar to importvar Fixed reading bad data on 86 hw init Fixed a bug loading 8xg files Fixed a bug in the symbol83p_t struct Fixed exporting groups

+++ WabbitStudio/Source/Wabbitemu/core/core.h
@@ -201,7 +201,7 @@ typedef struct memory_context {
 			BOOL flash_enabled : 1;
 			BOOL flash_disabled : 1;
 		};
-		uint16_t port24;
+		uint8_t port24;
 	};
 
 	int port27_remap_count;		// amount of 64 byte chunks remapped from RAM page 0 to bank 3
